backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=5.0.0
openai>=1.50.0
python-dotenv>=1.1.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.32.5
cryptography>=41.0.0
stripe>=7.0.0
alembic>=1.12.0

